Chore #16: rename designsystem -> component
