+++ feature/home/src/main/java/com/hankki/feature/home/HomeRoute.kt
@@ -39,12 +39,12 @@ import com.hankki.core.designsystem.component.bottomsheet.HankkiStoreJogboBottom
 import com.hankki.core.designsystem.component.bottomsheet.JogboItemEntity
 import com.hankki.core.designsystem.theme.Gray100
 import com.hankki.feature.home.MapConstants.DEFAULT_ZOOM
-import com.hankki.feature.home.designsystem.ChipState
-import com.hankki.feature.home.designsystem.DropdownFilterChip
-import com.hankki.feature.home.designsystem.HankkiTopBar
-import com.hankki.feature.home.designsystem.RepositionButton
-import com.hankki.feature.home.designsystem.RowFilterChip
-import com.hankki.feature.home.designsystem.StoreItem
+import com.hankki.feature.home.component.ChipState
+import com.hankki.feature.home.component.DropdownFilterChip
+import com.hankki.feature.home.component.HankkiTopBar
+import com.hankki.feature.home.component.RepositionButton
+import com.hankki.feature.home.component.RowFilterChip
+import com.hankki.feature.home.component.StoreItem
 import com.naver.maps.geometry.LatLng
 import com.naver.maps.map.CameraAnimation
 import com.naver.maps.map.CameraPosition
